Large refactoring to keep async/await calls in db.rs
With the new menu structure I was getting async recursion errors, so I added the "async-recursion" crate. That fixed the async recursion error. However, if you progressed into the multi-level menus like "Config/List/List Something" I was getting a hard crash of:

thread 'main' has overflowed its stack

So I unpicked all the async/await stuff to be blocked to run synchronously in the db layer. I may one day regret it but it's because of the db code and the "quirky nature" of async in Rust I ended up with async/await everywhere, when I didn't need it. This is a simple command line app, it doesn't need to run stuff asynchronously.